Make classloader customization easier to discover
If you encounter the problem of the two classloaders in Devtools, there's no hint that there are two classloaders and that some configuration changes might help you. You only get an IllegalArgumentException. Googling does not help because the exception is not mentioned in the docs. Futhermore, the docs sound like the problem can only occur in IDEs, which is not the case.