Adds support for Log4jBridgeHandler
Since version 2.15.0 `log4j-jul` contains a `Log4jBridgeHandler`, that
allows to forward JUL to Log4j 2.x and synchronizes the logger levels of
the two frameworks.

This PR adds support for the `Log4jBridgeHandler` and sets it as default
bridge handler for the Log4j 2.x stack and fallback handler for Logback.

It also removes `jul-to-slf4j` from the `spring-boot-starter-log4j2`,
hence leaving a pure Log4j 2.x stack.