Work around Gradle changing line endings during property expansion
There's a known issue [1] where property expansion changes the input
files line endings to the operating system's default. This causes
problems for us on Windows as the line endings become \r\n which
breaks our formatting checks.

This commit tunes the checkFormatMain task to exclude the generated
source files from checking. In their place, the original templates
are added. This ensures that the inputs are correctly formatted and,
therefore, that the output should be too (other than the line endings
on Windows).

Closes gh-30039

[1] https://github.com/gradle/gradle/issues/1151