Make the path of the config directory better understandable
The current documentation relates to the subdirectory for external
configurations as `/config`. Read as UNIX path, the leading slash
could be read as reference to the root directory `/`. Therefore
I would suggest to use `./config` to make it a little bit more
understandable.
This is of course not a major issue, but IMHO makes it a little
bit more understandable.