Fail when setReadTimeout on httpclient5 request factory
Prior to this commit, the `RestTemplateBuilder` would offer a generic
`setReadTimeout` method to configure the read timeout on the underlying
`ClientHttpRequestFactory`. This would be done in a reflective fashion,
considering that all implementations align with this behavior.

This option cannot be provided for HttpClient5 at the
`ClientHttpRequestFactory` level anymore, so this has been deprecated
 in Spring Framework 6.0 and will log a warning. In order to align with
 our existing behavior (throwing exceptions if the option cannot be set),
 this commit ensures that exceptions are also thrown is of the method is
 marked as deprecated.

32461