Adding missing BaggageTextMapPropagator for OTel W3C
without this change we're missing the BaggageTextMapPropagator for OTel. This means that we're not propagating remote-fields (only baggage via the 'baggage' field).

with this change we're adding the missing propagator as THE LAST entry in the composite TextMapPropagator. It has to be last cause with the latest Snapshots of Micrometer Tracing it will append the remote field baggage to existing baggage in the Context extracted via the W3CBaggagePropagator.