Add support for properties from file
With this change a property can be set using file content.
This works naming an environment variable with the UPPER_CASE
style, appending "_FILE" at it and setting the value to an absolute file path.
Thus, after checking for a system property and an environment
variable the properties loader checks for a environment variable
with this name and, if it exists try to read the content of the file
in the path declared.
This come very useful when working with docker secrets, so a property
can be overwritten using a secret and having an environment variable
with name "PROPERTY_NAME_FILE" with the path to the secret in it.