Adds CompositeSpanExporter
Micrometer Tracing comes with 3 generic interfaces, SpanExportingPredicate, SpanReporter and SpanFilter, thanks to which you can decide whether you want to export a span, how you want to report it and how to mutate it regardless of which tracer library you are using.

Without this change usage of these 3 interfaces is not possible for OpenTelemetry.

With this change we're putting all SpanExporters to the CompositeSpanExporter and apply the exporters only after predicates, filtering and reporting took place.