Consider properties on outer class in nested sliced tests
Prior to this commit, properties configured using the properties
attribute of a `@…Test` sliced test annotation would not be considered
when processed a `@Nested` test class. This would lead to the nested
class not reusing its outer class's application context due to the
two having different property configuration.

Fixes gh-33317