[PR requested change] "The property classes seem very flat. I think it would be good to group things together more."
I did this across all of the config props. Turned out to be a pretty big change
but well worth it. I do notice that some of the fields are out of alpha order but
the changes were already big enough that I think that can wait for a subsequent PR
if it is an issue.