Support @Order on [CommandLine|Application]Runner @Bean definitions
Prior to this commit, `@Order` annotation on `@Bean` method was not
considered for `CommandLineRunner` and `ApplicationRunner`.

This commit introduces a `Runner` marker interface and uses it to
retrieve the runner beans. As a result, it enables the use of `@Order`
annotations on `@Bean` methods for both `CommandLineRunner` and
`ApplicationRunner`.

Signed-off-by: Tadaya Tsuyukubo <tadaya@ttddyy.net>

See gh-37905